Report

+++ src/main/java/serverbyrtagihan/impl/CustomerImpl.java
@@ -30,13 +30,11 @@ import serverbyrtagihan.service.CustomerService;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
+import java.text.SimpleDateFormat;
 import java.util.*;
 
 @Service
-public class ProfileImpl implements CustomerService {
+public class CustomerImpl implements CustomerService {
     @Autowired
     CustomerRepository customerRepository;
 
@@ -55,6 +53,8 @@ public class ProfileImpl implements CustomerService {
     BillRepository billRepository;
     @Autowired
     TransactionRepository transactionRepository;
+    @Autowired
+    UserRepository userRepository;
 
     @Autowired
     PasswordEncoder encoder;
@@ -120,21 +120,18 @@ public class ProfileImpl implements CustomerService {
     public Map<Object, Object> login(LoginRequest loginRequest) {
         Customer customer = customerRepository.findByEmail(loginRequest.getEmail()).orElseThrow(() -> new NotFoundException("Username not found"));
         if (encoder.matches( loginRequest.getPassword(),customer.getPassword())) {
-
             Authentication authentication = authenticationManager.authenticate(
                     new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword()));
             SecurityContextHolder.getContext().setAuthentication(authentication);
             String jwt = jwtUtils.generateJwtToken(authentication);
-
-            LocalDateTime waktuSaatIni = LocalDateTime.now(ZoneId.of("Asia/Jakarta"));
-            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-            String waktuFormatted = waktuSaatIni.format(formatter);
-            customer.setLast_login(waktuFormatted);
+            customer.setLast_login(new Date());
             customerRepository.save(customer);
+            SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+            String formattedLastLogin = sdf.format(customer.getLast_login());
             Map<Object, Object> response = new HashMap<>();
             response.put("data", customer);
             response.put("token", jwt);
-            response.put("last_login", waktuFormatted);
+            response.put("last_login", formattedLastLogin);
             response.put("type_token", "Customer");
             return response;
         }
@@ -480,7 +477,7 @@ public class ProfileImpl implements CustomerService {
         String code = code();
         MimeMessage message = javaMailSender.createMimeMessage();
         MimeMessageHelper helper = new MimeMessageHelper(message, true);
-        if (customerRepository.existsByEmail(forGotPass.getEmail())) {
+        if (customerRepository.existsByEmail(forGotPass.getEmail()) || userRepository.existsByEmail(forGotPass.getEmail())) {
             helper.setTo(forGotPass.getEmail());
             helper.setSubject("Konfirmasi Reset Password");
             helper.setText("", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD XHTML 1.0 Transitional //EN\" \"http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd\">\n" +
@@ -732,9 +729,7 @@ public class ProfileImpl implements CustomerService {
                     "\n" +
                     "                              <div class=\"v-text-align\" align=\"left\">\n" +
                     "                                <!--[if mso]><table width=\"100%\"           cellpadding=\"0\" cellspacing=\"0\" border=\"0\" style=\"border-spacing: 0; border-collapse: collapse; mso-table-lspace:0pt; mso-table-rspace:0pt;font-family:arial,helvetica,sans-serif;\"><tr><td class=\"v-text-align v-button-colors\" style=\"font-family:arial,helvetica,sans-serif;\" align=\"left\"><v:roundrect xmlns:v=\"urn:schemas-microsoft-com:vml\" xmlns:w=\"urn:schemas-microsoft-com:office:word\" href=\"\" style=\"height:37px; v-text-anchor:middle; width:178px;\" arcsize=\"16%\" stroke=\"f\" fillcolor=\"#00afef\"><w:anchorlock/><center style=\"color:#FFFFFF;font-family:arial,helvetica,sans-serif;\"><![endif]-->\n" +
-                    "                                <a href=\"youtube.com\" target=\"_blank\" class=\"v-size-width v-button-colors v-border v-border-radius\" style=\"box-sizing: border-box;display: inline-block;font-family:arial,helvetica,sans-serif;text-decoration: none;-webkit-text-size-adjust: none;text-align: center;color: #FFFFFF; background-color: #00afef; border-radius: 6px; -webkit-border-radius: 6px; -moz-border-radius: 6px; width:auto; max-width:100%; overflow-wrap: break-word; word-break: break-word; word-wrap:break-word; mso-border-alt: none;\">\n" +
-                    "                                  <span class=\"v-line-height v-padding\" style=\"display:block;padding:10px 20px;line-height:120%;\"><span style=\"font-size: 14px; line-height: 16.8px;\">Setel Ulang Password</span></span>\n" +
-                    "                                </a>\n" +
+                    "                              \n" +
                     "                                <!--[if mso]></center></v:roundrect></td></tr></table><![endif]-->\n" +
                     "                              </div>\n" +
                     "\n" +
@@ -825,22 +820,24 @@ public class ProfileImpl implements CustomerService {
                     "</body>\n" +
                     "\n" +
                     "</html>");
-            Customer customer = customerRepository.findByEmail(forGotPass.getEmail()).get();
-            customer.setToken(code);
-            var checkingCode = getVerification.findByEmail(customer.getEmail());
-            if (getVerification.findByEmail(forGotPass.getEmail()).isPresent()) {
-                getVerification.deleteById(checkingCode.get().getId());
-                Reset_Password pass = new Reset_Password();
-                pass.setEmail(forGotPass.getEmail());
+             if (customerRepository.existsByEmail(forGotPass.getEmail())) {
+                Customer customer = customerRepository.findByEmail(forGotPass.getEmail()).get();
                 customer.setToken(code);
-                pass.setCode(code);
-                getVerification.save(pass);
-                customerRepository.save(customer);
-            } else {
-                Reset_Password pass = new Reset_Password();
-                pass.setEmail(forGotPass.getEmail());
-                pass.setCode(code);
-                getVerification.save(pass);
+                var checkingCode = getVerification.findByEmail(customer.getEmail());
+                if (getVerification.findByEmail(forGotPass.getEmail()).isPresent()) {
+                    getVerification.deleteById(checkingCode.get().getId());
+                    Reset_Password pass = new Reset_Password();
+                    pass.setEmail(forGotPass.getEmail());
+                    customer.setToken(code);
+                    pass.setCode(code);
+                    getVerification.save(pass);
+                    customerRepository.save(customer);
+                } else {
+                    Reset_Password pass = new Reset_Password();
+                    pass.setEmail(forGotPass.getEmail());
+                    pass.setCode(code);
+                    getVerification.save(pass);
+                }
             }
             javaMailSender.send(message);
         } else {
@@ -1206,17 +1203,17 @@ public class ProfileImpl implements CustomerService {
                     "\n" +
                     "</html>");
             if (customerRepository.existsByEmail(signupRequest.getEmail())) {
-                throw new NotFoundException("Kesalahan: Email telah digunakan!");
+                throw new NotFoundException(" Email telah digunakan!");
             }
             String UserEmail = signupRequest.getEmail().trim();
             boolean EmailIsNotValid = !UserEmail.matches("^(.+)@(\\S+)$");
             if (EmailIsNotValid) {
-                throw new BadRequestException("Kesalahan: Email tidak valid");
+                throw new BadRequestException(" Email tidak valid");
             }
             String UserPassword = signupRequest.getPassword().trim();
             boolean PasswordIsNotValid = !UserPassword.matches("^(?=.*[0-9])(?=.*[a-z])(?=\\S+$).{8,20}");
             if (PasswordIsNotValid) {
-                throw new BadRequestException("Kesalahan: Password tidak valid");
+                throw new BadRequestException(" Password tidak valid");
             }
             // Create new user's account
             Customer admin = new Customer();
@@ -1226,6 +1223,7 @@ public class ProfileImpl implements CustomerService {
             admin.setHp(signupRequest.getHp());
             admin.setName(signupRequest.getName());
             admin.setAddress(signupRequest.getAddress());
+            admin.setOrganization_id(null);
             admin.setToken("Kosong");
             javaMailSender.send(message);
             return customerRepository.save(admin);
